+++ src/const.js
@@ -1,0 +1,5 @@
+const TYPES = [`Taxi`, `Bus`, `Train`, `Ship`, `Transport`, `Drive`, `Flight`, `Check-in`, `Sightseeng`, `Restaurant`];
+const NAMES = [`Amsterdam`, `Geneva`, `Chamonix`];
+const OFFERS = [`Add luggage`, `Switch to comfort class`, `Add meal`, `Choose seats`, `Travel by train`];
+
+export {TYPES, NAMES, OFFERS};
